We created a new class of concrete wall by analogy with a brick one. We added the remaining map objects - water, forest, ice and eagle. Added the ability to draw sprites on different layers. We added one more object - the map border and adjusted the window size to the loaded map size. Added spawn locations for players and enemies to the level description. Added a Timer class to implement time-limited animations. Added a physics engine class for the movement of game objects. Added the ability to shoot. Animation of projectiles has been created.

+++ src/Renderer/SpriteAnimator.h
@@ -1,20 +1,22 @@
-#pragma once 
+#pragma once
 
 #include <memory>
-namespace RenderEngine{
+
+namespace RenderEngine {
 
     class Sprite;
-    class SpriteAnimator{
+
+    class SpriteAnimator {
     public:
         SpriteAnimator(std::shared_ptr<Sprite> pSprite);
-        size_t getCurrentFrame() const {return m_currentFrame;}
-        void update(const uint64_t delta);
+        size_t getCurrentFrame() const { return m_currentFrame; }
+        void update(const double delta);
 
     private:
         std::shared_ptr<Sprite> m_pSprite;
         size_t m_currentFrame;
-        uint64_t m_currentFrameDuration;
-        uint64_t m_currentAnimationTime;
+        double m_currentFrameDuration;
+        double m_currentAnimationTime;
     };
 
 }
